Python GUI Drill 3 Step 123
For this drill, you will need to gather all of the drills that you have previously completed for this course and write a new script that will use each of the concepts that you had previously used to complete this drill assignment.

Your new script will need to provide the user with a graphical user interface that includes two buttons allowing the user to browse their own system and select a source directory and a destination directory. Your script should also show those selected directory paths in their own corresponding text fields.

Next, your script will need to provide a button for the user to execute a function that should iterate through the source directory, checking for the existence of any files that end with a “.txt” file extension and if so, cut the qualifying files and paste them within the selected destination directory.

Finally, your script will need to record the file names that were moved and their corresponding modified time-stamps into a database and print out those text files and their modified time-stamps to the console.